+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to this project will be documented in this file.
 - Notify when trying to send to a terminal that hasn't been started yet.
 - Available text decorators by name can be configured in `config.text_decorators`.
 - `.select()` can take a single function instead of a table of callbacks.
+- `.select()` to fast-forward if single terminal and single default callback
 
 ## 0.5.0 - 2025-09-27